requirements.txt
requests==2.31.0
python-dotenv==1.0.0
gitpython==3.1.40
loguru==0.7.2
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
httpx==0.25.2 